new registration tests
